+++ NewSuperMarioBrosPC/Leaf.h
@@ -13,7 +13,7 @@ public:
 	static const string OBJECT_NAME;
 	static const int HEIGHT;
 	static const int WIDTH;
-	Leaf(int X, int Y, int Width, int Height, int vx, int vy, int vx_last, int ax, int ay, CSprite * Sprite);
+	Leaf(int X, int Y, int Width, int Height, float vx, float vy, float vx_last, float ax, float ay, CSprite * Sprite);
 	string getName()override;
 	void onCollision(Object * ob, int dir);
 };
